refactor hooks and modal to use index.tsx
