+++ frontend/src/components/ResultsView/ItemMenu/PersonCard/personCard.component.tsx
@@ -1,7 +1,6 @@
 import { ChangeEvent, useEffect, useRef, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { didMount } from "../../../../hooks/didMount";
-import { didClickAway } from "../../../../hooks/didClickAway";
+import { didClickAway, didMount } from "../../../../hooks";
 import { AppStore, Item, Person } from "../../../../models/custom-models";
 import { editPersonName } from "../../../../redux/calculation/calculation-actions";
 import { ItemRow } from "./ItemRow";
