+++ frontend/src/components/ResultsView/ItemMenu/itemMenu.component.tsx
@@ -7,8 +7,7 @@ import {
   editEventTitle,
   recalculateEvent,
 } from "../../../redux/calculation/calculation-actions";
-import { didMount } from "../../../hooks/didMount";
-import { didClickAway } from "../../../hooks/didClickAway";
+import { didClickAway, didMount } from "../../../hooks";
 import { dismissToast, errorToast } from "../../shared/toasts/toasts";
 import { validString } from "../../../utilities/sanitize";
 import { ERROR_INPUT_EVENT } from "../../../utilities/variables";
