+++ frontend/src/components/shared/modal/index.tsx
@@ -1,0 +1,1 @@
+export { Modal } from "./modal.component";
