--- frontend/src/hooks/didMount/index.tsx
@@ -1,1 +1,0 @@
-export { didMount } from "./didMount.hook";
