+++ frontend/src/hooks/index.tsx
@@ -1,0 +1,2 @@
+export { didMount } from "./didMount/didMount.hook";
+export { didClickAway } from "./didClickAway/didClickAway.hook";
